Move camera spawning/tracking to a `CameraPlugin`
## This Commit

Extracts a plugin dedicated to spawning the camera and moving it around
to follow the player.

## Why?

When I extracted `camera_follow_player` into the `PlayerPlugin` I was
actually hesitant and thought that went better in a `CameraPlugin`. Then
the `PlayerPlugin` would move the `Player` and the `CameraPlugin` would
move the `Camera`. Then I was reviewing [this Bevy Jam submission][0]
and saw [this][1] and thought, "Okay, let's pull the trigger on that and
see how it goes".

## Note

In that repo their `follow_camera` utility eventually got slurped up
into a [`Controllable` component][2] so maybe we get there eventually.

[0]: https://github.com/Louis-Tarvin/Elemental-Sorcerer
[1]: https://github.com/Louis-Tarvin/Elemental-Sorcerer/commit/3b2d535f768f4167dcc798a2f85631b1f574abaf#diff-e73f90ab0f0491e6368341bc6018aad7721ea34aff137e157619a6b75f5b4529R1
[2]: https://github.com/Louis-Tarvin/Elemental-Sorcerer/blob/0007fb23aa25bf9b6d7a5587897739336aaa40ac/src/input.rs#L18
